Fixed the wrong dialog location when user has multiple monitors
Dialog windows would be displayed on the wrong screen location if the
user has a system with multple monitors. Because we don't have some
wxDisplay functions available in Erlang I did a workaround by using the
minimal coordinate of the window instead of zero. That is necessário
only if the display in use is placed on the left or top side of the
primary monitor - in these cases the coordinates has negative value.

NOTE: Fixed the wrong dialog location when user has multiple monitors.
Thanks to suzuki.