Workd on Sculpt code in order to fix some issues
1) Fixed an issue which was rollbacking the #st{} in sculpt mode;
2) By addapting the previous fix, it was forced Sculpt to init in
   shade mode. If it was started with a geometry window in Wire frame
   mode.

The problem stated after the UI migration. Before the main menu was disabled
in Sculpt mode making the Geometry window "modal" and that couldn't happens.
It now can deal with that as well as the synchronization between geometry
windows allow us to see the changes applied to the object it doesn't matter
if it was done in sculpt window or any other geometry - before that was
impossible.
Originally, the sculpt states was storing a #st{} structure
three times. After this change, it is now keeping only two of them.
I avoided to make any drastic change to the code to avoid instability, but
maybe in the future we could to evaluate a way to store only one #st{}
structre ('stw' - the internal work one) as well as remove any event that
became obsolete now.
I tested the save/restore windows; the File menu options with the plugin
active and no problem was verified.

NOTE: - Fixed the problem that was causing geometry state rollback after
save a project in Sculpt mode. Thanks to tkbd.
NOTE: - By starting Sculpt with a geometry window in Wire mode would make
the sculpt never get back to Shading mode. Thanks to Hank.