Fixed the restored size for external windows (#308)
Fixed the restored size for external windows

I just found the issue related to the floating windows (free) which was
making them grow each time Wings3D was started. We are storing the window's
size (the whole window), but we are recovering that information to the
window client's size which doesn't include the title and border size.

Note: Floating windows was growing each time Wings3D was started.