Added the PLY importer to the Wings3D project (#313)
Added the PLY importer to the Wings3D project

PLY importer was not working for most the cases. It was fixed the code by
adding some ways to handle the '.ply' file compatibility as well as handle
the exception caused by the use of a binary format which is still not
supported and may be implemented further.

* Included the alpha channel for the vertex color if present.

NOTE: Added the PLY importer to the Wings3D project.