Added a scale system for ground grid visualization
This scale system was implemented because the auto-grid was causing redraw
issues (system freezing) when a large object was imported - as reported at
the forum. It will not fill the viewport with the grey lines side-by-side
making a "plane" as before.

It was also made a small adjustment to Y coordinate where the letters were
printed out since the "Y" used to be cut in the top-half on screen.

OBS: The clip region will need to be adjusted by the user when a large
     scene is load.