Fixed the crash caused by Tighten in AutoUV window
The problem was affecting only the Tighten command for the AutoUV window and
it was related to an empty list of vertex passed to drag module.
As I could check, it was introduced after the changes made to wings_drag module
in versin 2.1.6.

NOTE: Fixed the crash caused by Tighten in AutoUV window. Thanks to Hank.