AutoUV window been shown in the same display where the segmenting window is
It was asked by the user OXO to make possible to continue the UVmapping in
the same monitor (for multiple monitors setup). This way, if the user start
to segmenting the mesh in another monitor than the one in which the main
window is, the next AutoUV window will be displayed in the same monitor.

NOTE: UV mapping windows are displayed in the same monitor as the first was.
Thanks OXO.