Fixed hard crash caused by malformed faces
It was provided a model by Hank in which after evaluation was found a face
with five vertex, but no area. That was resulting in a zero lenght normal.
The change is intended to avoid Wings3D get a hard crash by a user fault.

NOTE: Fixed hard crash caused by malformed faces. Thanks Hank.