Fixed the crash caused by Tighten in AutoUV window (#328)
* Fixed the crash caused by Tighten in AutoUV window

The problem was affecting only the Tighten command for the AutoUV window and
it was related to an empty list of vertex passed to drag module.
As I could check, it was introduced after the changes made to wings_drag module
in version 2.1.6.

Also fixes 80be5b so revert that commit:

* Revert "Bevel command was crashing when a vertex which dived a edge"
This reverts commit 80be5b2a71b912a8a06bef1523fbe349dc545539.

NOTE: Fixed the crash caused by Tighten in AutoUV window. Thanks to Hank.