Boolean fixes
Restart more often, restart every time we touch a face already
changed.

It is hard to get the correct EPSILON :-)

We cannot have a to small limit because it will create small (flat)
triangles which can not be intersected in the coming passes.

Also dissolve reworks faces, be sure that temp edges are not in the
border edges when we cleanup afterwards.

NOTE: Worked some more on the boolean operations.