Added option to create non modal dialog/window
It was no necessary to change relevant part of the code, but add funtions
to support the needs to make it work.
The forms layout are defined just like we do for modal dialogs and the
creation funcion requires a hook that will handle the changes for each field's
value, then it get notified after any field has been changed - there is no
need for define hook per field/control.
The returned value must be stored in the window #state{} since it represents
the pair {ets,keys} - just like used in wings_dialog.

The 'preview' event is fired to the form window and as well as for dialog it
means some field's value was changed. For the case we need to take all field's
values in this momment there is the funciont form_result for that.