Fixed a crash caused by wings_view during vectors input
When jumping to a saved view during a vector input operation and the
current saved view was the same as the current one, the code was returning
a 'keep' value instead of the non changed #st{} - which was the expected
value by the wings_vec module.

NOTE: Changing to a saved view in the middle of a vector input operation
was causing Wings3D to crash. Thanks to Hank.