Workaround to a hard crash when closing a non registere window
It was obseved a crash in wings_frame module when handling the #wxClose event
which was unexpected and means an unhandled #wx{obj} is firing this event.
As a workaroound we are going to check for that and display some information
that would help us to know what is that object and how to permanently to fix
the crash cause.

Reported here: http://www.wings3d.com/forum/showthread.php?tid=2733

NOTE: Wings3D was crashing unexpected in module wings_frame. It was add a
workaround and some information to help us to find the cause. Thanks to  Hank