Delay callback invocation
Otherwise slider invokes the callbacks which may take some time to render
before the function is returned and the process state is update.
If paint event comes before the callbacks are done a redraw will be done
on the old state causing the slider pointer to jump between states.

On mouse move set state and return, delay callback invocation.
Then both callbacks and paint event will be handled with the same state.