Fix almost opauge transperent (smooth) rendering
Transparent faces are drawn with depth-write off and without sorting
which cause many render articfacts.

One major artifact is for the opaque parts, which should be drawn
with depth-write on.

Since an earlier fix all textures with an alpha channel are considered
transparent even if they only contain a=1 everywhere.

Fix this by rendering to the depthbuffer if alpha is greater than
threshold. Otherwise we skip it. Also use a polygon-offset and depth
test LEQUAL to avoid z-buffer fighting.