Fixed the wrong location for uv segmentation submenu.
For some strange reasion the Y coordinate returned by mouse_get_state/1 when
a submenu needed to be shown was getting a value greater than ~4000000 instead
of a lower negative value for the layout where the second monitor is up or left
to the first one.
But, for some reason, converting it to Client and back to Screen fix the
coordinate value and the menu is shown in the right place.

NOTE: The submenu in 'AutoUV Segmentating' was shown in the wrong display for
multiple monitors layout. Thanks to OXO.