Pluging that add a Isometric view option
It will not reset the current view, but just adjust the camera location and
mode (Orthogonal) and it will be applied for the current geometry window (in
case we use a layout with others than the default one).

This plugin is intended to help modelers to see how their project will looks
like in games which uses the Isometric view.
It was a need by the user SiriusCG which we needed to trick Wings3D to be
able to reach the expected effect.