Changed the splash window with a user image
The idea is to open a new challenge at the forum calling the users to create
a splash image. The community will vote and the winner image is added to the
splash window in the next Wings3D release.

To allow that it was changed the way the dialog is built and drawn to allow
us draw text over an image.
It was also set the focus to the window when it is called from Help->About.