Fixed the plugin/patch installer
The plugin/patch installer was not allowing to pack patches into a .tar file.

OBS: It's important to remember that a '.tar' file must to contain or patches or
plugins, since its priority is for plugins (wpc_*) and mixing them will result
in patches files being unpacked under plugins folder.