Add Drag'n'Drop of image from outliner to autouv
But only change the background, don't update the material as was done
previously.

Since now you might want to see other types of textures than color when lining
up your uv-coords but not make that a base-color texture.

We could pop up a menu and ask and update the material but that is
a future improvment.

Also we might need that when we render textures, i.e. what type of
texture should we add it as.

NOTE:
Allow image drag'n'drop from outliner to autouv to change
background image in autouv.