Save Selected was saving extra images
The changes was tested with objects without material; objects with
material and image; objects with material and no image. All of project
file with a lot of images loaded (internal and external).

NOTE: The option "File->Save Selected" was saving images not used by the
selected object(s); Thanks Hank