Fixed an hotkey isseue related to main menu shortcuts
If we were bind a hotkey that was own by a main menu item Wings3D was asking
for replace it, but was not really doing that. Also, a try to set a hotkey
back to the main menu was not updating it as well as removing the previous
command that was using it.

NOTE: Binding a hotkey used by main menu was not removing it. Thanks yschaeff