Add a mousewheel event
Refactor mousewheel events and add possibility to get HORIZONTAL
scroll events.

Enabled when wx have it, the horizontal events must be handled inside
a callback currently, so that is done and sent to the wings process.

NOTE:
Enable horizontal scroll to rotate view, improves laptop usage
without mouse.  Can be used with options see:
"Preferences/Camera/Wheel Pans and Rotates"
and it's options.