Context menu on Geometry window always fits inside the display
The context menu on geometry window can get the bottom portion hidden by the
task bar. Also, if the user hit RMB with the menu shown the menu can be cut
once the new top location is the mouse position.

It was used wsDisplay information to compute the correct location of a menu
on a display where the mouse was used to invoke it. The getClientArea function
takes in account the task bar if it's present. That ensure we don't get any
portion of the menu hidden by it.

NOTE: Context menu on Geometry can get options hidden. Thanks to nigec.