Added the toggle background option back to the AutoUV window
This option was moved from AutoUV window to Show menu  during the migration
to wx library. It's very handy we have it very close to our work on the UV
mesh. To get it available the AutoUV window must to be active and sometimes
moving the mouse from AutoUV window to Show menu results in other window
grab the focus and the option becomes not acessible.

NOTE: Added the toggle background option to the AutoUV window.