menu fix(es) incl prev commits
Previous menus commits was not intended to be merged to maint yet, sigh.

Anyway these commits adds an (almost) transparent overlay window,
to catch clicks outside of wings windows, to be able to close the menu
if clicked outside or if ESC pressed.

The previous solution grabbed mouse to do that but caused that we lost
tooltips for the menu items.

If I make the overlay window completly transparent on windows and mac
no events are received so it must have "some" color.