Fixed bad behaviour of screenshot option
The screenshot command was always getting the image size from the main geometry
window (by considering a layout with multiple ones)- wings_wm:this() was always
referencing 'geom' window. But the copied buffer was the last one drawn (maybe).
To ensure we were copying the proper one it was added a new function to
wings_wm to allow us require a specific geometry window to be drawn and then
we can access the correct OpenGL buffer.

When a detached geometry window is active it loses the focus when we use the
main menu. So, in this case the main geometry window will be used for capture.

It was also removed the option to capture the entire window since it was not
working well since we started to use the wx library (requested by Dan).

NOTE: Fixed the screenshot command which was only capturing the main
geometry window and cutting it on top.