Fixed the missing right message in the information line
I'm not sure this fix is appropriated, but it was needed in order to ensure the
wm_focus data used in wings_wm:message_event(redraw) get the currect focused
window instead of 'undefined' and this way the status bar can be updated.

NOTE: The right message used to show extra information to the user actions
was missing with the previous message kept static. Thanks to Sevendy3