Fixed the missing fields on Material tab
It was missing some fields in Material dialog: the 'Fake Shadows' for Rough
Glass and 'Emit Ligth', 'Oren-Nayar' and 'Fresnel Effect' for Shinny Diffuse.

That happened due to deprecated fields that were removed. Dynamics dialogs in
Wings3D (the ones we show/hide fields) requires we to hide some fields
in order the dialog doesn't reserve a high frame to show all the added fields.
As some fields were removed the height reserved was not enough to display all
of them.

We need to be sure to keep visible  all the fiedls that can be shown at one
specific moment in which the dialog will have highest high.