The changes to makefile allow us to build ligigl from MSys
It were made changes in makefile to the parameters xxFLAGS replacing the
slash(/) by the dash(-) to avoid the compiler to fail in MSys environment.
This change will not cause any problem to its use in WSL environment.

In the .bashrc we then can set the variable UseMCL like this below. The
explicit cl.exe path addition is required because MSys will put its system
folder at the beggining of $PATH and will cause the MS link not be used.

MCL="$(which cl.exe | grep -i -Po 'cl.exe')"
if test "x$MCL" = "x"; then
  echo "Linux Compiler to be used."
  export CC=gcc
  export OPENCL_DIR=/mingw64
else
  echo "Microsoft Compiler to be used."
  export UseMCL=true
  export CC=cl.exe
  export OPENCL_DIR="'/C/Program Files (x86)/Microsoft Visual Studio/2019/Community/
                      VC/Tools/MSVC/14.26.28801'"
  export PATH="/c/Program Files (x86)/Microsoft Visual Studio/2019/Community/VC/Tools/
               MSVC/14.26.28801/bin/Hostx64/$BV":$PATH
fi