Added AutoUV shader texture preview
As more I was playing with this stuff, creating new shaders, as more I missed
an preview option. I heard this complain from other users too. So, I decided
to try one approach.

It's hard to make some actions work in the "right way" when some code is
working on different process. So, I did my best to not mess up with the
source code.

Please take a look, try it and see if it could be added to the main source
with any changes you judge to be necessary.