Fixed crash caused by garbage left when dissolving faces
It was noticed that by using commands like 'Select Similar Normals' the
faces marked as hole was also selected and by dissolving the selection
the #we{}.holes was not updated to remove these faces.

It was opted to keep the holes or hidden faces select since for some commands
it can be of interest move these faces too. This is noticeable with hidden
faces when becoming visible will be close to the ones moved previously.