Fixed collada importer error
It was noticed collada may generate files with NaN for float numbers. As
checked, others 3D apps uses to automatically to convert them to 0.0, so
it was changed the wings_util:string_to_float/1 in order to fix that.
This will prevent any other importer to break Wings3D in the future.

NOTE: Fixed the collada inporter erro caused by NaN values. Thanks to tkbd