Added option to AutoUV exports UVs
That can be useful to someone export a open mesh to be used in other
purposes than 3D modelling, like to cut cardboard, paper or metal.
The idea comes from the discussion in this thread:
http://www.wings3d.com/forum/showthread.php?tid=2965

NOTE: Added an option to AutoUV to export the UV mesh as .eps/.svg files.