Checking for a previous libigl build on 'priv' folder
During the build/rebuld/test process of Wings3D the 'libigl' is rebuilt
all the time if we are using make from Wings3D's root directory and/or switching
between branches. That's very annoying and it stoles from us some precious
time.
Also, if no internet connection is available the dependencies checking
seems to hang - which should also not be a problem if we already have a built
version of the library available. To update it we can just run 'make clean'.