Fixed the issue reported by Travis as asked by dgud
The problem was related to a empty parameter I was send to a function I created
in a previous fix.
As dgud also asked to avoid to use pass a list of lists as parameter for
format commands I fixed the issu by replacing that list by a format string
and a list of paramters.

p.s.I also adjusted the output alignment to xml for the 'mesh' session, then I
was able to better read the result of the changes.