Added a dialog to STL exporter (#424)
* Added a dialog to STL exporter

It was added a simplified export dialog including the option for scale
and subdivide the scene. It also includes a preset table with a list of
some knowing slicer softwares and its measurement references related to
1 STL unit - which should help the users to understand its relationship
with the Wings3D unit (1WU = 1STL unit); a drop down list for unit
convertion (1WU to 1STL/1mm/1cm/1inch), the conversion scale and the
object scale.
The combination of the first two generates the value for the conversion
scale which multiplied by the object scale results in a update for the
'Export scale' text box.

* Selecting 'default Wings3D settings' reset the obj_scale and conv_scale to 1.0

* After we make a couple of tests using many slicers it was noticed that all
of them translate 1STL unit to 1mm, which is different from the initial
information we got. That happened because those slicers were trying to make
the very tinny test object (1x1x1mm) in something bigger.
Again, thanks to oort for help me with the tests.

NOTE: Added a dialog to STL exporter allowing users to set the scale.
Thanks to dawntreader, imagine and Mert HANCIOGLU(instagram) for provide
us informations about the slicers measures.