Release dialog grab when window is closed
Closing a window with a non-modal dialog (i.e. preview dialogs) open,
caused the mouse grab to stay forever and wings was dead.

Fix by cancel the dialog when the window is closed, this will leave
the previewed data because we can't send the message to the window
which have been closed. Can't think of a general workaround, the
window might not be a geom window (autouv for example).