A fix for the wrong mouse wheel behaviour in slicer controls
Since the previous fix to make the values increase or decrease in accordly
when using mouse wheel in a textbox or a slider, the overal behavior was changed
with the wheel going up to decrease the value and going down do increate it.

That became a change in the way the textbox controls has been working on Wings3D
for long time - which bring us a new problem.

It was found that this wrong behaviour on mouse wheel over the slider has been
reported time to time, but not fix or even a comment about that.
It works correctly if the wxSL_VERTICAL style is used (which can be combined
with the style wxSL_INVERSE).

So, this commit is a workaround that fix the problem by handling by ourselves
the mouse_wheel event.