Added option to avoid shaders to use the preview sphere in dialog
It was added a new paramter to the wpc_***.auv that let the code know it
doesn't requires/produce a texture preview.

It was also made some changes to prevent crashes if a shader settings previously
stored in the preferences and load in a new project references an image that is
no long available or a shader that was removed. In these situations the stored
information is removed from the layer/slot before the Options dialog be shwon.