Added option to avoid shaders to use the preview sphere in dialog
It was added a new paramter to the wpc_***.auv that let the code know it
doesn't requires/produce a texture preview.
It was also made some changes to prevent crashes if a shader settings previously
stored in the preferences and load in a new project references an image that is
no long available. Unfortunately, we don't have a way to know if between those
recovered parameters some of them is related to an image until we try to bind
them to the shader uniforms.
So, I splited out some functions to be reused and then we are now checking if
an image is available otherwise we replace it by the default AuvBG.