Added support for fast loading environment maps
It was added code to store the pre-computed environment maps data in the cache
folder. The files get the '.emc' extension which stands for 'Environment Maps
Cache'. For the current settings (MMs) the file size is 9MB.
In case of fail to reading the file it is deleted and the code to create the
environment maps on-the-fly will run.
Even the default 'grandcanyon' image will be cached, so the loading time and
the switching process to other environment maps start to happening in no time.

The fake_envmap process stays the same, but it could be redesigned to use the
same file cache approach with the file being packed with Wings3D.