Small adjustments to the geometry visualisation
- Decreased the near clipping value from 0.01 to 0.001;
This limitation has been commented time to time. I already experienced that
sometimes. So, by changing it we open the possibility to work correctly with
detailed object which has small pieces that need attention.
- Added the possibility to draw the ground grid for scales 0.1, 0.01 and 0.001;
The grid for these scales was missing.
- Adjusted the function bind for drawing the main axes. The exact 0.0 value
barely is reached and sometimes they are hidden by the grid line.

NOTE:
- Decreased the near clipping value from 0.01 to 0.001;
- The ground grid is also now drawn for scales as 0.1, 0.01 and 0.001;