Fixed a crash caused by checking image name
Sometimes when the image list has an image with name is 'none' Wings3D is going
to hard crash. That was introduced in the current version when it was tried to
improve the evaluation for an already loaded image.

NOTE: Fixed a crash caused by checking image name. Thanks to rgcotl (at Discord)