Moved the material properties for plugins to another window
The change was made to speed up the access to the Material dialog which
became too slow since the UI was migrated to use wx library.
When an user choose a render plugin to change its setting that will be
assumed as his/her prefered render engine (which usually is). So, in the
next time the material dialog be accessed it will automatically set to
use that one.

p.s.: I tried to minimize the changes and keep the material data consistent with
any Wings3D version.

NOTE: The material properties for plugins are now accessed by choosing the
plugin name in a drop-down list and clicking on the "Edit..." button.