Update view auto-rotate fps
Render as much as possible and sync with gl:finish(),
i.e. waiting on monitor sync instead waiting of an erlang timer
that could be late.

Means that we block the wx-thread for 15ms while waiting for sync,
so we might have a lot of mouse events to handle before rendering
the next frame, i.e event handling might be unresponsive on slow
machines and large models.