Do not spawn process in render dialog
The crash was caused by two processes calling OpenGL simultanously,
which caused havoc.

One solution is to remove the 'preview' handling in the top dialog,
which will make the dialog modal and opengl calls from the second
dialog will be ok.

But we can keep the 'preview' functionality by sending a callback to
the autouv window instead, since when the dialog has a preview window
it is not modal and can handle events, and thus creating the second
dialog which contains opengl code, but is now blocking autouv
(and wings main process).

- spawn(fun() ->
-    	%% Need open dialog in dialog from another process
-    	wx:set_env(Env),
-    	option_dialog(Id, Fields, Renderers, Shaders),
-            wings_wm:psend(send_once, dialog_blanket, preview)
-       end)
+ Dlg = fun() ->
+          option_dialog(Id, Fields, Renderers, Shaders),
+          wings_wm:psend(send_once, dialog_blanket, preview),
+          keep
+       end,
+ wings_wm:psend(Me, {callback, Dlg}),

Also contains some indentation fixes.

NOTE: Fixed autouv crash in create_texture dialogs.