Added option to rotate the environment map
It's a missing option (in my opinion) since mostly 3D apps which works with
PBR visualisation allow us to rotate the environment map around the static
object to give us the possibility to evaluate how the object is lit by
different angles.

NOTE: Added option to rotate the environment map.