Making context menu always fits inside the display
It was noticed the context menus has stopped to be shown close the mouse on a
multiple display setup. This issue was workaround in the past in PR#392. But,
it seems wxWidgets has fixed the issue with the mouse location in this kind
of setup bring back this kind of issue.

NOTE: Fixed misplaced context menu in a multiple display setup.