Keep track of reparent
After a reparent that somehow involves the wxGLCanvas, i.e. a direct
reparent of the wxGLCanvas or a grandparent of the wxGLCanvas use
wxGLCanvas::CreateSurface() and recreate the wxGLContext::new( with previous context )

After a reparent that does not involve the wxGLCanvas or its
grandparents, do NOT call wxGLCanvas::CreateSurface() because that
will fail (return false) for some unknown reason and all opengl calls
will return errors.