Fixed a crash when a context menu is active and we close its parent window
This error was noticed in the AutoUV Segmentation window when the context
menu was active and the use clicked in the [X] to close its parent window.
Ref.: http://www.wings3d.com/forum/showthread.php?tid=3093

NOTE: Fixed the cause of the crash when closing AutoUV Segmentation window
with the context menu active. Thanks to markie.