Workarounded a hard crash when starting Wings3D in aPC with an Intel GPU
That is not clear if it's a GPU driver or wxWidgets issue.
After a couple of tests playing around with the 'Intel Graphics Control Panel'
settings I found out the issue was caused by the MSAA setting. By switching it
from 'Turn Off' to 'Use Application Settings' fixed the problem.

Since it seems MSAA is turned off by default in some systems and it's hard to
the user to figure out it's causing the Wings3D failing to start, I'm proposing
to disable the use of MSAA for Intel GPUs.

NOTE: This fix a hard crash on starting Wings3D in which a PC's setup that
uses an Intel as main GPU.