Fixed an issue to Connect RMB option when only one edge was selected
It was noticed if we have only one edge selected and the Connect RMB command
is activated the it was entering in slide mode even without a valid edge to
move. That could trick the user.

NOTE: Fixed the Connect RMB command that was entering in slide mode after no
connection has been done.