Disabling MSAA feature in case of hard crash on Wings3D start (#527)
* Disabling MSAA feature in case of hard crash on Wings3D start

That is not clear if it's a GPU driver or wxWidgets issue. After a couple of
tests playing around with the 'Intel Graphics Control Panel' settings I found
out the issue was caused by the MSAA setting. By switching it from 'Turn Off'
to 'Use Application Settings' fixed the problem.

Because it seems MSAA is turned off by default in some systems and because it's
hard to the user to figure out it's causing the Wings3D failing to start, I'm
proposing a similar behavior used for OpenCL validation. In this case, after
Wings3D fail if the user try to run it once again then Wings3D will start with
MSAA feature disabled.
In order to try enable it again the user will need to remove the msaa.txt file
from the cache folder.
Since the validation is made on init/0 a variable in preferences is going to
help to avoid to check that file when we call window/4 to create a new GL window.

NOTE: Fixed a crash which happened during start on some Intel GPUs when MSAA was 
disabled in a PC setup.