Fixed the impossibility to bind DELETE key after it has been unbound
There are a couple of default hotkey set to use DELETE key. So, it was
allowed to use it to be bind to any command, as well as INSERT, but taking
care about the combination with CTRL which are used to invoke bind/unbind
commands.

NOTE:
Fixed the bind DELETE key after it has been unbound. Thanks to ptoing.