Fixed a division by zero crash in AutoUV when very tiny islands exists
tkbd reported and provided a model which very tiny islands were present
in AutoUV and by applying Scale->Uniform, Scale->Max or Remapping was
causing Wings3D to crash by badarith - because of their zero size area
which was causing a division by zero.
The changes cover the three situations which the crash could happen.

It was added a function (nonzero/1) to wings_util.erl and a constant
EPSILON to wings.hrl.

NOTE:
Fixed a division by zero crash in AutoUV when very tiny islands
exists. Thanks tkbd