Fixed the menu location in order to fit on the display it is activated
This issue was added when I remove a old workaround for multiple display
setup and didn't pay attention that it was also making the menu window
to fit on a display. I revert the commit and rewrote it.

It's not perfect for setups where the displays have different resolution.
In that situation if a window is shared by multiple displays in a location
where a window is scaled the correct dimensions can not be safely evaluated
and then we can still get the popup menu cropped.

NOTE: Fixed the popup menu location that can be shown out of screen.
Thanks to Xavier