Updated table control to allow us handle its content
It was made changes to the code in order to allow us to replace the table
content. This way we can implemente addition, update and deletion of items
in a table list.

It was also disable the title highlight stuff since we don't handle its
event in order to promote ordering which can cause confusion to the user.