Fixed the menu location in order to fit on the display it is activated (#583)
* Fixed the menu location in order to fit on the display it is activated

This issue was added when I remove a old workaround for multiple display
setup and didn't pay attention that it was also making the menu window
to fit on a display. I revert the commit and rewrote it.

It's not perfect for setups where the displays have different resolution.
In that situation if a window is shared by multiple displays in a location
where a window is scaled the correct dimensions can not be safely evaluated
and then we can still get the popup menu cropped.


* Fixed the menu selection that was letting garbage
This issue was introduced by the fix for another menu issue that was crashing
Wings3D when closing a window with the context menu active (commit: #be68730)

The fix redraw the entire menu unselected before it be displayed. It was also
needed to find for the window of the submenu item because it's not stored in
the object property of menudata.

NOTE: Fixed popup menu location that could be shown out of screen. Thanks to Xavier

 Fixed the menu selection that was showing garbage. Thanks to sciroccorics