Move hard-edge from vertex-normal calcs from e3d to wings
To be able to improve finding hardedges from vertex normals, we
need to postpone the calculation til we have a #we{} to work on,
so we can fold over faces and/or vertices.

So include the normal information in wings_we_build.
If there are no hardedges and and vertex normals try to find an
approximation of the hardedges from the vertex normals.

I do that in 2 steps first add edges that are clearly hard
or delete cleary not hard.

Pick out vertices which are neither.

Second pass loop around vertices and find hard edges.