Fixed dialog placement on a multiple-display setup.
A Discord user, Aleks M, reported an issue with his setup, which included two
displays (one horizontal and one vertical). When the dialog was invoked with
the mouse near the left border of the vertical display, the dialog was cut off.

To address this, I updated the code to evaluate which display the mouse was on,
and used its geometry to compute the optimal position for showing the dialog.
This solution worked well; however, in cases where the display uses scaling
(via the OS), the dialog's position can still vary slightly, sometimes ending
up farther from the mouse pointer than expected.

I also addressed an issue with the option to display centered dialogs.
On multiple displays, the dialog could sometimes be positioned across the frame
of the displays, resulting in only half of it being visible on each display,
which was quite annoying.