Updated the circularise command to better handle mirrored selections
It was changed the code in order to evaluate the linked edges selection
and see if they lay on the mirror face boundaries. In this case, the
linked edges are handled as they were edge loops and we get the command
working as it was applied in a non mirrored object.
In case other linked edges exists and not laying on mirror face boundaries
then the error message caused by mixed mode is displayed, just as it uses
to be for non mirrored objects.
NOTE:
- Updated the circularise command to better handle mirrored selections;