Workaround to fix text colour not changing in status bar
It was noticed that the text color in the status bar doesn't change according
to the theme settings. Looking for info I found that it's an know issue in
wxWidgets which setting the color of native controls is not always possible.

Considering we use the wxFLAT style, drawing the status bar ourselves is the
easiest and most effective way to work around this issue. I included a grip
icon similar to the Windows version, since no system resource is available
for use (as with other system icons).

NOTE:
Fixed a theme issue that was preventing the status bar text from using
its setting in the theme color. Thanks to Nova711 (on Discord).