Added an option to apply ambient occlusion (AO) blending with vertex colors
To enable a different action via the application menu, the CTRL modifier key
was used to trigger a secondary option for the AO command. This is a useful
feature for those who paint vertex objects for use in games.

NOTE:
Added an option to apply AO blending with vertex colors. Thanks to boggers
and DevEarley for the suggestion