New AutoUV Scale->Normalize RMB option
It's intended to normalize a set of charts relative to a reference one.

This is option is useful in cases where we need to remap a portion of an
already unwrapped object, or when someone unwraps a composite object in
parts. In these situations, we want the new charts to remain consistent in
scale with the previously unwrapped ones. To achieve this, we simply select
one of them as a reference.

NOTE:
New AutoUV option to Normalize a set of charts using another as reference.
It's accessed via RMB option. Thanks to envelupo and Vaughan for the suggestion